refactor(asr): isolate packed attention utilities
